accomodation_id error fixes

+++ src/main/java/com/example/szallas/repository/RatingRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface RatingRepository extends JpaRepository<Rating, Long> {
-    List<Rating> findByAccommodationId(Long accommodationId);
+public interface RatingRepository extends JpaRepository<Rating, Integer> {
+    List<Rating> findByAccommodationId(Integer accommodationId);
 }
